rename module names in game_interface
