+++ xlive/Blam/Engine/Simulation/game_interface/simulation_game_objects.h
