+++ xlive/Project_Cartographer.vcxproj
@@ -756,13 +756,13 @@
     <ClInclude Include="Blam\Engine\scenario\scenario_interpolators.h" />
     <ClInclude Include="Blam\Engine\scenario\scenario_kill_trigger_volumes.h" />
     <ClInclude Include="Blam\Engine\shell\windows\crash_report_window\crash_report_window.h" />
-    <ClInclude Include="Blam\Engine\Simulation\GameInterface\SimulationGameEntities.h" />
-    <ClInclude Include="Blam\Engine\Simulation\GameInterface\SimulationGameObjects.h" />
-    <ClInclude Include="Blam\Engine\Simulation\GameInterface\SimulationGameUnits.h" />
-    <ClInclude Include="Blam\Engine\Simulation\GameInterface\simulation_game_engine_player.h" />
+    <ClInclude Include="Blam\Engine\Simulation\game_interface\simulation_game_entities.h" />
+    <ClInclude Include="Blam\Engine\Simulation\game_interface\simulation_game_objects.h" />
+    <ClInclude Include="Blam\Engine\Simulation\game_interface\simulation_game_units.h" />
+    <ClInclude Include="Blam\Engine\Simulation\game_interface\simulation_game_engine_player.h" />
     <ClInclude Include="Blam\Engine\Simulation\MachineID.h" />
     <ClInclude Include="Blam\Engine\Simulation\SimulationActors.h" />
-    <ClInclude Include="Blam\Engine\Simulation\GameInterface\SimulationGameEvents.h" />
+    <ClInclude Include="Blam\Engine\Simulation\game_interface\simulation_game_events.h" />
     <ClInclude Include="Blam\Engine\Simulation\SimulationPlayers.h" />
     <ClInclude Include="Blam\Engine\Simulation\SimulationTypeCollection.h" />
     <ClInclude Include="Blam\Engine\Simulation\SimulationView.h" />
